+++ playwright/tests/review.spec.ts
@@ -1,10 +1,16 @@
 import { test, expect } from "@playwright/test";
 import { ReviewPage } from "../pageobjects/review.page";
-import { TEST_PATRON_INFO } from "../utils/constants";
+import {
+  TEST_PATRON_INFO,
+  USERNAME_AVAILABLE_MESSAGE,
+  USERNAME_UNAVAILABLE_MESSAGE,
+} from "../utils/constants";
+import { mockUsernameApi } from "../utils/mock-api";
 
 test.beforeEach(async ({ page }) => {
   await page.goto("/library-card/review?newCard=true");
 });
+
 test.describe("displays elements on review page", () => {
   test("displays headings", async ({ page }) => {
     const reviewPage = new ReviewPage(page);
@@ -21,6 +27,7 @@ test.describe("displays elements on review page", () => {
     await expect(reviewPage.emailHeading).toBeVisible();
     await expect(reviewPage.receiveInfoHeading).toBeVisible();
   });
+
   test("displays Address section headings", async ({ page }) => {
     const reviewPage = new ReviewPage(page);
     await expect(reviewPage.addressHeading).toBeVisible();
@@ -68,5 +75,41 @@ test.describe("edits patron information on review page", () => {
     );
     await expect(reviewPage.emailInput).toHaveValue(TEST_PATRON_INFO.email);
     await expect(reviewPage.receiveInfoCheckbox).not.toBeChecked();
+  });
+
+  test("displays Create Your Account section headings", async ({ page }) => {
+    const reviewPage = new ReviewPage(page);
+    await expect(reviewPage.createYourAccountHeading).toBeVisible();
+    await expect(reviewPage.usernameHeading).toBeVisible();
+    await expect(reviewPage.passwordHeading).toBeVisible();
+    await expect(reviewPage.showPassword).toBeVisible();
+    await expect(reviewPage.homeLibraryHeading).toBeVisible();
+    await expect(reviewPage.ebranchValue).toBeVisible();
+    await expect(reviewPage.createYourAccountEditButton).toBeVisible();
+    await expect(reviewPage.submitButton).toBeVisible();
+  });
+});
+
+test.describe("mock API responses on Review page", () => {
+  test("displays username available message", async ({ page }) => {
+    // mock the API call for username availability
+    await mockUsernameApi(page, USERNAME_AVAILABLE_MESSAGE);
+
+    const reviewPage = new ReviewPage(page);
+    await reviewPage.createYourAccountEditButton.click();
+    await reviewPage.usernameInput.fill("AvailableUsername");
+    await reviewPage.availableUsernameButton.click();
+    await expect(reviewPage.availableUsernameMessage).toBeVisible();
+  });
+
+  test("displays username unavailable error message", async ({ page }) => {
+    // mock the API call for username unavailability
+    await mockUsernameApi(page, USERNAME_UNAVAILABLE_MESSAGE);
+
+    const reviewPage = new ReviewPage(page);
+    await reviewPage.createYourAccountEditButton.click();
+    await reviewPage.usernameInput.fill("UnavailableUsername");
+    await reviewPage.availableUsernameButton.click();
+    await expect(reviewPage.unavailableUsernameError).toBeVisible();
   });
 });