+++ playwright/utils/mock-api.ts
@@ -1,4 +1,6 @@
-export async function mockUsernameApi(page, message: string) {
+import { Page } from "@playwright/test";
+
+export async function mockUsernameApi(page: Page, message: string) {
   await page.route("**/library-card/api/username", async (route) => {
     await route.fulfill({
       status: 200,